added feign client to denormalize image in post and reel service as well as added controller and service to manage counter in profile

+++ src/main/java/com/example/Social/profile/dto/fetchProfile.java
@@ -18,11 +18,11 @@ public class fetchProfile {
 
     private Boolean privateAcc;
 
-    private  int followerCount;
+    private  long followerCount;
 
-    private  int followingCount;
+    private  long followingCount;
 
-    private  int friendsCount;
+    private  long friendsCount;
